requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
pydantic==2.5.0
librosa>=0.10.0
numpy>=1.21.0
aiofiles==23.2.0
onnxruntime
openai-whisper>=20231117